submenu delete and validasi done

+++ public/private_file/assets/js/master/preferences/submenu.js
@@ -33,20 +33,73 @@ $(document).ready(function() {
     $('#table').on('click', '#delete', function (e) {
         e.preventDefault();
         let value = $(this).data('value')
-        $.ajax({
-            url:'/api/v1/submenu/delete',
-            data:{
-                value : value
+        SweetQuestions({
+            title : 'Apakah anda yakin?',
+            subtitle : 'Apakah anda ingin menghapus data section ini?',
+            buttonConfirm : 'Yes',
+            buttonDeny: 'No',
+            confirm : 'ajax',
+            deny : {
+                icon:'error',
+                title : 'Gagal menghapus'
             },
-            type:'DELETE',
-            headers:{
-                'X-CSRF-TOKEN' : csrftoken
+            ajax : {
+                url:'/api/v1/section/delete',
+                data:{
+                    value : value
+                },
+                type:'DELETE',
+                headers:{
+                    'X-CSRF-TOKEN' : csrftoken
+                },
+                success:res=>{
+                    SweetAlert(res)
+                    RefreshTable('table')
+                    value_checkbox = []
+                },
+                error:err=>{
+                    SweetAlert({status:'error', message:err.responseJSON.message})
+                    value_checkbox = []
+                }
+            }
+        })
+    })
+
+    $('#deleteArray').on('click', function (e) {
+        e.preventDefault();
+        if (value_checkbox.length < 1) {
+            Swal.fire('Perhatian', 'Pilih salah satu!','warning')
+            return 0;
+        }
+        SweetQuestions({
+            title : 'Apakah anda yakin?',
+            subtitle : 'Apakah anda ingin menghapus data submenu ini?',
+            buttonConfirm : 'Yes',
+            buttonDeny: 'No',
+            confirm : 'ajax',
+            deny : {
+                icon:'error',
+                title : 'Gagal menghapus'
             },
-            success:res=>{
-                SweetAlert(res)
-                RefreshTable('table')
-            },
-            error:err=>console.log(err)
+            ajax : {
+                url:'/api/v1/submenu/delete',
+                data:{
+                    value : value_checkbox
+                },
+                type:'DELETE',
+                headers:{
+                    'X-CSRF-TOKEN' : csrftoken
+                },
+                success:res=>{
+                    SweetAlert(res)
+                    RefreshTable('table')
+                    value_checkbox = []
+                },
+                error:err=>{
+                    SweetAlert({status:'error', message:err.responseJSON.message})
+                    value_checkbox = []
+                }
+            }
         })
     })
